# services-api/app/Models/Department.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\HasMany;

class Department extends Model
{
    use HasFactory;
    protected $fillable = [
        'nameDep','description','manager'
    ];

    public function services(): HasMany
    {
        return $this->hasMany(Service::class);
    }
}