# services-api/app/Models/Service.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\BelongsTo;
use Illuminate\Database\Eloquent\Relations\HasMany;

/**
 * Class Service
 *
 * @package App\Models
 * @property int $id
 * @property string $name
 * @property string $description
 * @property float $price
 * @property bool $is_active
 * @property string $location
 * @property string $contact_email
 * @property string $contact_phone
 */
class Service extends Model
{
    use HasFactory;

    protected $fillable = [
        'name','description','price','is_active','location','contact_email','contact_phone'
    ];

    public function department(): BelongsTo
    {
        return $this->belongsTo(Department::class);
    }

    public function employee(): BelongsTo
    {
        return $this->belongsTo(Employee::class);
    }
}